added speed/pitch touchscreen menu to control it

+++ include/settings.h
@@ -5,11 +5,11 @@
 inline constexpr std::string_view SETTINGS_PATH = "sdmc:/3ds/Canary/settings.ini";
 
 struct Settings {
-    int volumePercent = 100;        // 0-VOLUME_MAX_PERCENT
-    int brightness = 3;             // 1-5 levels
-    int seekSeconds = 10;           // seek duration in seconds
-    int speedPercent = 100;         // playback speed, SPEED_MIN_PERCENT-SPEED_MAX_PERCENT
-    int pitchSemitones = 0;         // pitch shift, PITCH_MIN_SEMITONES-PITCH_MAX_SEMITONES
+    int volumePercent = 100;      // 0-VOLUME_MAX_PERCENT
+    int brightness = 3;           // 1-5 levels
+    int seekSeconds = 10;         // seek duration in seconds
+    int speedPercent = 100;       // playback speed x100, SPEED_MIN_PERCENT-SPEED_MAX_PERCENT
+    float pitchSemitones = 0.0f;  // pitch shift, PITCH_MIN_SEMITONES-PITCH_MAX_SEMITONES, 0.1 steps
     bool linkedSpeedPitch = false;  // when true, pitch follows speed
     std::string startPath = "sdmc:/Music/";
     bool lockToStartPath = true;  // prevent navigating above start path
